Fixed auto loading zshrc with tmux and i3
I was having issues getting i3 and tmux working out the gate,
so it would be helpful to keep the auto-loading for those
programs turned off until I can get into the new system
and make sure everything was configured properly.

To turn them back on, I'd advise to copy and paste the commented
code from zshrc and paste it into the ~/.zshrc instead of in the
git repo.  Just in case you want to make new uploads, this wont
get disturbed on each individual machine.